// src/app/models/emppriorexperiencerequest.ts
export class EmployeePriorExperienceRequest {
    id: number;
    employeeId: number;
    departmentId: number;
    designationId: number;
    office: string;
    range: any;
    dateOfJoining: any;
    dateOfLeaving: any;
    experience: any;
    isActive: string;
    modifiedBy: number;
    createdBy: number;
}